rename + changelog + cleanup
