+++ src/app/editor/components/templateExamples.tsx
